Only execute yarn install if not in dry mode

+++ .ivy/raise-deps.sh
@@ -3,4 +3,6 @@ set -e
 
 sed -i -E "s/(\"@axonivy[^\"]*\"): \"[^\"]*\"/\1: \"~${1/SNAPSHOT/next}\"/" packages/*/package.json
 yarn update:axonivy:next
-yarn install --ignore-scripts
+if [ "$DRY_RUN" = false ]; then
+  yarn install --ignore-scripts
+fi
